% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{sfaira_ref}
\alias{sfaira_ref}
\title{hand sfaira from specified basilisk environment}
\usage{
sfaira_ref()
}
\description{
hand sfaira from specified basilisk environment
}
\note{
sfaira 0.3.12 munged, tensorflow 2.8.0, protobuf 3.20.0
}
