% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{sfaira_ui}
\alias{sfaira_ui}
\title{hand back ui.UserInterface from sfaira from specified basilisk environment}
\usage{
sfaira_ui(ref)
}
\description{
hand back ui.UserInterface from sfaira from specified basilisk environment
}
\note{
sfaira 0.3.12 munged, tensorflow 2.8.0, protobuf 3.20.0
}
